+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+## Added
+- Added `is_state_tcp`, `is_state_udp`, and `is_state_closed` to the `Common` trait.
+
 ## [0.2.0] - 2021-02-10
 ### Added
 - Added feature flags to pass through to w5500-ll.